Move out all functions
